Fixes restart sesssion bug and refactors controls container
